feat: Enhance hotel management features
- Added new status options for hotels (active, suspended, closed) in data.js.
- Introduced amenities list for hotels in data.js.
- Implemented HotelStatus component for managing hotel status.
- Updated HotelEditForm to include status and amenities management.
- Added API endpoint to fetch hotel details by ID.
- Enhanced hotel deletion confirmation in HotelCards component.
- Improved UI for adding hotels with a new button design.
- Updated hotel management page to navigate to add hotel form.
- Refactored hotelController to handle cover image uploads and address parsing.
- Added CSS styles for new components and improved existing styles.

+++ client/src/components/sideBar/hotelCards/HotelCards.jsx
@@ -13,15 +13,18 @@ export const HotelCards = ({ hotel, hotels, setHotels }) => {
 
   const DeleteHotel = async (id) => {
     try {
-      const response = await deleteHotel(id);
-      if (response?.data?.success) {
-        const hotelList = [];
-        for (let i = 0; i < hotels.length; i++) {
-          if (hotels[i]._id !== id) {
-            hotelList.push(hotels[i]);
+      const deleteChoice = confirm("Ary you sure you want to delete this?");
+      if (deleteChoice) {
+        const response = await deleteHotel(id);
+        if (response?.data?.success) {
+          const hotelList = [];
+          for (let i = 0; i < hotels.length; i++) {
+            if (hotels[i]._id !== id) {
+              hotelList.push(hotels[i]);
+            }
           }
+          setHotels(hotelList);
         }
-        setHotels(hotelList);
       }
     } catch (error) {
       console.error(error);
@@ -83,7 +86,7 @@ export const HotelCards = ({ hotel, hotels, setHotels }) => {
             className="btn-icon"
             title="Edit Hotel"
             onClick={() => {
-              navigate("/admin/hotel-edit");
+              navigate(`/admin/hotel-edit/${hotel._id}`);
             }}
           >
             <FontAwesomeIcon icon={faEdit} />
